+++ resources/views/gudangBarangJadi/requestPotong/index.blade.php
@@ -51,9 +51,11 @@
                 <div class="col-12">
                     <div class="card">       
                         <div class="card-header">
-                            <h3 class="card-title" style="width: 100%">
-                                <a href="{{ route('GBarangJadi.requestPotong.create') }}" class='btn btn-info btn-flat-right'>Tambah Request</a>
-                            </h3>
+                            @if (\Auth::user()->roleId == 1 || \Auth::user()->roleId == 5 || \Auth::user()->roleId == 8 || \Auth::user()->roleId == 37)
+                                <h3 class="card-title" style="width: 100%">
+                                    <a href="{{ route('GBarangJadi.requestPotong.create') }}" class='btn btn-info btn-flat-right'>Tambah Request</a>
+                                </h3>
+                            @endif
                         </div>                 
                         <div class="card-body">
                             <table id="example2" class="table table-bordered dataTables_scrollBody" style="width: 100%">
@@ -79,9 +81,12 @@
                                             </td>
                                             <td>
                                                 <a href="{{ route('GBarangJadi.requestPotong.detail', [$req->id]) }}" class='btn btn-warning'><i class="fas fa-list-ul" style="font-size: 14px"></i></a>
-                                                <a href="{{ route('GBarangJadi.requestPotong.update', [$req->id])}}" class='btn btn-success'><i class="fas fa-pencil-alt" style="font-size: 14px"></i></a>
-                                                @if ($req->statusDiterima == 0)
-                                                    <button type="button" data-toggle="modal" requestId='{{ $req->id }}' data-target="#DeleteModal" id="modalDelete" onclick='deleteData("{{ $req->id }}")' class='btn btn-danger delete'><i class="fas fa-trash" style="font-size: 14px"></i></a>        
+                                                
+                                                @if (\Auth::user()->roleId == 1 || \Auth::user()->roleId == 5 || \Auth::user()->roleId == 8)
+                                                    <a href="{{ route('GBarangJadi.requestPotong.update', [$req->id])}}" class='btn btn-success'><i class="fas fa-pencil-alt" style="font-size: 14px"></i></a>
+                                                    @if ($req->statusDiterima == 0)
+                                                        <button type="button" data-toggle="modal" requestId='{{ $req->id }}' data-target="#DeleteModal" id="modalDelete" onclick='deleteData("{{ $req->id }}")' class='btn btn-danger delete'><i class="fas fa-trash" style="font-size: 14px"></i></a>        
+                                                    @endif
                                                 @endif
                                             </td>
                                         </tr>
